Gets drive data on login. Won't open past drives until drives are recieved from the sheet.
